Use conditional compilation for platform specific functions
This ensures that the crate is buildable on all platforms. If building on non-x86_64, not all functions will be present.
