+++ src/instructions/segmentation.rs
@@ -9,16 +9,6 @@ use structures::gdt::SegmentSelector;
 /// and return value on the stack and use lretq
 /// to reload cs and continue at 1:.
 pub unsafe fn set_cs(sel: SegmentSelector) {
-    #[cfg(target_arch = "x86")]
-    #[inline(always)]
-    unsafe fn inner(sel: SegmentSelector) {
-        asm!("pushl $0; \
-              pushl $$1f; \
-              lretl; \
-              1:" :: "ri" (u64::from(sel.0)) : "rax" "memory");
-    }
-
-    #[cfg(target_arch = "x86_64")]
     #[inline(always)]
     unsafe fn inner(sel: SegmentSelector) {
         asm!("pushq $0; \
