+++ src/instructions/tables.rs
@@ -2,16 +2,7 @@
 
 use structures::gdt::SegmentSelector;
 
-/// A struct describing a pointer to a descriptor table (GDT / IDT).
-/// This is in a format suitable for giving to 'lgdt' or 'lidt'.
-#[derive(Debug, Clone, Copy)]
-#[repr(C, packed)]
-pub struct DescriptorTablePointer {
-    /// Size of the DT.
-    pub limit: u16,
-    /// Pointer to the memory region containing the DT.
-    pub base: u64,
-}
+pub use structures::DescriptorTablePointer;
 
 /// Load GDT table.
 pub unsafe fn lgdt(gdt: &DescriptorTablePointer) {
