+++ src/structures/idt.rs
@@ -429,6 +429,7 @@ impl InterruptDescriptorTable {
     }
 
     /// Loads the IDT in the CPU using the `lidt` command.
+    #[cfg(target_arch = "x86_64")]
     pub fn load(&'static self) {
         use core::mem::size_of;
         use instructions::tables::{lidt, DescriptorTablePointer};
@@ -550,6 +551,7 @@ impl<F> Entry<F> {
     ///
     /// The function returns a mutable reference to the entry's options that allows
     /// further customization.
+    #[cfg(target_arch = "x86_64")]
     fn set_handler_addr(&mut self, addr: u64) -> &mut EntryOptions {
         use instructions::segmentation;
 
@@ -566,6 +568,7 @@ impl<F> Entry<F> {
 
 macro_rules! impl_set_handler_fn {
     ($h:ty) => {
+        #[cfg(target_arch = "x86_64")]
         impl Entry<$h> {
             /// Set the handler function for the IDT entry and sets the present bit.
             ///
