+++ src/structures/mod.rs
@@ -3,4 +3,16 @@
 pub mod gdt;
 pub mod idt;
 pub mod paging;
+pub mod port;
 pub mod tss;
+
+/// A struct describing a pointer to a descriptor table (GDT / IDT).
+/// This is in a format suitable for giving to 'lgdt' or 'lidt'.
+#[derive(Debug, Clone, Copy)]
+#[repr(C, packed)]
+pub struct DescriptorTablePointer {
+    /// Size of the DT.
+    pub limit: u16,
+    /// Pointer to the memory region containing the DT.
+    pub base: u64,
+}
